Implement functionality for non-existent hookableMethods.
So far the hookz module looks good enough to let people
play with it. It is important to have some demo application
that uses this module, so it is easier for people to
understand what we are trying to achieve here.

Is there any other benefit from Backbone.Events besides defining
the events of the object by simply prefixing the method?

Also there are some features that are not implemented yet.
The comparison with the ProcessWire's Hooks is needed. Particularly
the ability to run the hooks before the event is not there yet. Which
is one of killer features of PW Hooks.

You can only connect the dots at the end. Meanwhile create whatever is
beauty to you.